debate_engine/requirements.txt
streamlit
pandas
numpy
plotly
yfinance
groq
python-dotenv
gTTS
fastapi
uvicorn
